Refactored to separate the controller methods serving the HTML pages from the RESTful controller methods, and removed the HTML methods from the Swagger documentation
