+++ src/main/java/application/SwaggerConfig.java
@@ -1,6 +1,7 @@
 package application;
 
 import com.google.common.base.Predicates;
+import io.swagger.annotations.Api;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import springfox.documentation.builders.ApiInfoBuilder;
@@ -19,7 +20,7 @@ public class SwaggerConfig {
     public Docket api() {
         return new Docket(DocumentationType.SWAGGER_2)
                 .select()
-                .apis(RequestHandlerSelectors.any())
+                .apis(RequestHandlerSelectors.withClassAnnotation(Api.class))
                 .paths(Predicates.not(PathSelectors.regex("/error.*")))
                 .build();
     }
